+++ src/index.ts
@@ -1,4 +1,4 @@
 export { isCompatibleType } from "./is-compatible-type.ts";
 export { createIsSameTypeFn, isSameType } from "./is-same-type.ts";
 export { defineCompareRule, isSameTypePresetRules } from "./rules.ts";
-export type { IsSameTypeOptions } from "./utils.ts";
+export type * from "./types.ts";
